+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for attributes in `Newline`.
+- `AttributedStringBuilder` protocol with default implementations of basic attribute modifiers.
+
+### Changed
+
+- `AString` and `Newline` now conform to the `AttributedStringBuilder` protocol.
+
 ## [0.0.2] - 29.01.20
 
 ### Added